requirements.txt
aiofiles==22.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.0
aiosignal==1.3.2
alembic==1.14.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.13.0
bleach==6.2.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
croniter==1.3.7
cryptography==42.0.0
cycler==0.12.1
dask==2025.1.0
datadog==0.44.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
entrypoints==0.4
executing==2.2.0
Faker==4.14.0
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.55.8
fqdn==1.5.1
freezegun==1.2.2
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.41
great-expectations==0.18.12
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.10
inflection==0.5.1
ipykernel==6.15.0
ipython==8.10.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==1.1.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==7.4.4
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.0.13
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
ldap3==2.9.1
Levenshtein==0.26.1
locket==1.0.0
mage-ai==0.9.75
makefun==1.15.6
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.1
multidict==6.1.0
mysql-connector-python==9.2.0
narwhals==1.24.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
newrelic==8.8.0
notebook==7.0.8
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.3.6
plotly==6.0.0
polars==0.19.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==1.59.0
Pygments==2.19.1
PyJWT==2.6.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==3.2.1
python-Levenshtein==0.26.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
redis==5.0.8
referencing==0.36.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.17.17
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==1.19.1
setuptools==70.0.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
sqlglot==26.3.9
sqlglotrs==0.3.14
stack-data==0.6.3
terminado==0.17.1
text-unidecode==1.3
thefuzz==0.19.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tomlkit==0.13.2
toolz==1.0.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.9.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.10.0
tzdata==2025.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.3.0
watchdog==4.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
